Complete GoRouter configuration and navigation

+++ ecoscan_ai/lib/presentation/screens/category_breakdown_screen.dart
@@ -57,7 +57,7 @@ class CategoryBreakdownScreen extends StatelessWidget {
           map[cat]!.red++;
           break;
       }
-      map[cat]!.totalScore += r.analysis.overallScore;
+      map[cat]!.totalScore += r.analysis.overallScore as int;
     }
     // Sort by total descending
     final sorted = Map.fromEntries(map.entries.toList()
